#CAP-11 blake3 and sha256 compare benchmark

+++ pallets/difficulty/src/lib.rs
@@ -120,7 +120,7 @@ pub mod pallet {
 			let mut current_height = CurrentHeight::<T>::get().unwrap_or(0u32);
 
 			// panic if current height pointer is over the boundray.
-			if current_height >= DIFFICULTY_ADJUST_WINDOW {
+			if current_height >= DIFFICULTY_ADJUST_WINDOW as u32{
 				panic!("current height pointer out of bounds!");
 			}
 			// It's time to adjust the difficulty
@@ -158,9 +158,9 @@ pub mod pallet {
 				// Difficulty adjustment and storage
 				let difficulty = Self::difficulty().unwrap_or(DIFFICULTY_DEFAULT) as i128 + adj_ts;
 				let mut difficulty_final = MIN_DIFFICULTY;
-				if difficulty < MIN_DIFFICULTY {
+				if difficulty < MIN_DIFFICULTY as i128{
 					difficulty_final = MIN_DIFFICULTY;
-				} else if difficulty > MAX_DIFFICULTY {
+				} else if difficulty > MAX_DIFFICULTY as i128 {
 					difficulty_final = MAX_DIFFICULTY;
 				} else {
 					difficulty_final = difficulty as u128;
